added GoRouter function
