+++ android/app/build.gradle
@@ -31,6 +31,8 @@ android {
         targetSdk = flutter.targetSdkVersion
         versionCode = flutter.versionCode
         versionName = flutter.versionName
+
+        minSdkVersion 23
     }
 
     buildTypes {
